chore(jans-lock): Further improved project structure
